Aggiunto al metodo di CUDA i parametri per dire la dimensione reale dell'input dei dati e la dimensione reale della dimensione dell'output.

+++ BeliefPropagationCUDA/myLib.h
@@ -58,7 +58,7 @@ extern "C"
 	    return ((x&(x-1))==0);
 	}
 	
-	unsigned int nextPow2(unsigned int x)
+	size_t nextPow2(size_t x)
 	{
 	    --x;
 	    x |= x >> 1;
